requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
numpy==1.24.3
pandas==2.1.1
scikit-learn==1.3.2
xgboost==2.0.2
yfinance==0.2.32
requests==2.31.0
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.0
gunicorn==21.2.0
huggingface-hub==0.17.3